requirements.txt
# ============================================================================
# 수어 번역기 시퀀스 관리 시스템 의존성 패키지 목록
# ============================================================================

# 필수 기본 패키지
python-dotenv>=1.0.0         # 환경변수 관리 (.env 파일)
numpy>=1.24.0                # 수치 계산 및 배열 처리

# PyQt5 GUI 패키지 (메인 UI)
PyQt5>=5.15.0                # GUI 프레임워크
PyQt5-Qt5>=5.15.0            # Qt5 바이너리
PyQt5-sip>=12.11.0           # SIP 바인딩

# OpenCV 컴퓨터 비전 (카메라 및 이미지 처리)
opencv-python>=4.8.0         # OpenCV 메인 패키지
opencv-contrib-python>=4.8.0 # OpenCV 추가 기능

# YOLO 모델 (수어 인식)
ultralytics>=8.0.0           # YOLOv8 모델 라이브러리
torch>=2.0.0                 # PyTorch (YOLO 백엔드)
torchvision>=0.15.0          # PyTorch 비전 라이브러리

# OpenAI 패키지 (문장 생성)
openai>=1.0.0                # OpenAI API 클라이언트

# Google Cloud 패키지 (TTS, STT)
google-cloud-texttospeech>=2.14.0    # Google Cloud Text-to-Speech
google-cloud-speech>=2.20.0          # Google Cloud Speech-to-Text
google-auth>=2.17.0                  # Google 인증
google-api-core>=2.11.0              # Google API 코어

# 오디오 처리 패키지
pygame>=2.5.0                # 오디오 재생 (TTS)
SpeechRecognition>=3.10.0    # 음성 인식 라이브러리
pyaudio>=0.2.11              # 오디오 입력/출력 (마이크 사용)

# 유틸리티 패키지
requests>=2.28.0             # HTTP 요청
urllib3>=1.26.0              # URL 처리
certifi>=2022.12.7           # SSL 인증서
pillow>=10.0.0               # 이미지 처리 (PIL)

# ============================================================================
# 운영체제별 설치 가이드
# ============================================================================

# Windows 설치:
# 1. Python 3.8+ 설치
# 2. pip install -r requirements.txt
# 3. 만약 PyQt5 설치 오류 시: pip install PyQt5 --force-reinstall
# 4. 만약 PyAudio 오류 시: pip install pipwin && pipwin install pyaudio

# macOS 설치:
# 1. Python 3.8+ 설치
# 2. brew install portaudio  # PyAudio용
# 3. pip install -r requirements.txt

# Linux (Ubuntu/Debian) 설치:
# 1. sudo apt-get update
# 2. sudo apt-get install python3-dev python3-pip
# 3. sudo apt-get install portaudio19-dev python3-pyaudio  # PyAudio용
# 4. sudo apt-get install python3-pyqt5 python3-pyqt5.qtmultimedia  # PyQt5용
# 5. sudo apt-get install libgl1-mesa-glx  # OpenCV용
# 6. pip install -r requirements.txt

# ============================================================================
# 주요 패키지별 기능
# ============================================================================

# PyQt5: 메인 GUI 인터페이스
# - 카메라 화면 표시
# - 텍스트 박스 (단어 누적 표시)
# - 수어하기/말하기 버튼
# - 실시간 UI 업데이트

# OpenCV: 카메라 및 이미지 처리
# - 웹캠 영상 캡처
# - 프레임 리사이즈
# - 가이드 박스 그리기
# - BGR ↔ RGB 변환

# Ultralytics (YOLO): 수어 인식 모델
# - best_1.pt 모델 로드
# - 실시간 수어 감지
# - 신뢰도 기반 필터링
# - 15개 클래스 분류

# OpenAI: 자연어 문장 생성
# - 수어 단어 → 자연스러운 문장 변환
# - GPT-3.5-turbo 모델 사용
# - 캐시 시스템으로 속도 최적화

# Google Cloud TTS: 음성 합성
# - 한국어 음성 합성
# - 다양한 음성 옵션
# - MP3 포맷 출력
# - pygame으로 재생

# Google Cloud STT: 음성 인식
# - 한국어 음성 → 텍스트 변환
# - 실시간 녹음 및 변환
# - 토글 방식 녹음

# ============================================================================
# 시스템 요구사항
# ============================================================================

# Python 버전: 3.8 이상 (권장: 3.9+)
# RAM: 최소 8GB (YOLO 모델 로드용)
# GPU: CUDA 지원 GPU 권장 (CPU도 가능하지만 느림)
# 디스크: 최소 2GB (모델 파일 및 캐시용)
# 카메라: USB 웹캠 또는 내장 카메라
# 마이크: 음성 인식용 (STT 기능 사용 시)
# 스피커: 음성 출력용 (TTS 기능 사용 시)

# ============================================================================
# 환경변수 설정 (.env 파일)
# ============================================================================

# 다음 내용으로 .env 파일 생성:
# OPENAI_API_KEY=your_openai_api_key_here
# GOOGLE_APPLICATION_CREDENTIALS=path/to/google-credentials.json

# ============================================================================
# 모델 파일 필요
# ============================================================================

# models/best_1.pt 파일이 필요합니다.
# - YOLO 수어 인식 모델
# - 15개 클래스 (구급차1/3, 구급차2/3, 구급차3/3, 학교, 쓰러지다1/2, 쓰러지다2/2, 아프다, 가다, 나, 사람1/2, 사람2/2, 빨리, 병원, 구조, 리셋)

# ============================================================================
# 설치 확인 방법
# ============================================================================

# 1. 패키지 설치 확인:
# python -c "import PyQt5; import cv2; import ultralytics; import openai; print('모든 패키지 설치 완료')"

# 2. 카메라 확인:
# python -c "import cv2; cap = cv2.VideoCapture(0); print('카메라 사용 가능' if cap.isOpened() else '카메라 오류')"

# 3. YOLO 모델 확인:
# python -c "from ultralytics import YOLO; model = YOLO('models/best_1.pt'); print('모델 로드 성공')"

# ============================================================================
# 문제 해결
# ============================================================================

# PyQt5 설치 오류:
# - pip install --upgrade pip setuptools wheel
# - pip install PyQt5 --no-cache-dir

# OpenCV 오류:
# - Linux: sudo apt-get install libglib2.0-0 libsm6 libxext6 libxrender-dev libgl1-mesa-glx
# - pip install opencv-python-headless  # GUI 없는 버전

# YOLO/Torch GPU 사용:
# - CUDA 11.8+: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
# - CUDA 12.1+: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121

# PyAudio 설치 오류:
# - Windows: pip install pipwin && pipwin install pyaudio
# - macOS: brew install portaudio && pip install pyaudio
# - Linux: sudo apt-get install portaudio19-dev && pip install pyaudio

# Google Cloud 인증 오류:
# - 서비스 계정 키 JSON 파일 다운로드
# - GOOGLE_APPLICATION_CREDENTIALS 환경변수 설정
# - 또는 gcloud auth application-default login

# ============================================================================
# 개발 및 디버깅 (선택사항)
# ============================================================================

# pytest>=7.0.0                # 테스트 프레임워크
# black>=23.0.0                # 코드 포매터
# flake8>=6.0.0                # 코드 린터
# jupyter>=1.0.0               # 노트북 (데이터 분석용)

# ============================================================================
# 라이센스 정보
# ============================================================================

# 이 프로젝트는 다음 오픈소스 라이브러리를 사용합니다:
# - PyQt5 (GPL v3)
# - OpenCV (Apache 2.0)
# - Ultralytics YOLO (AGPL-3.0)
# - 기타 패키지들은 각각의 라이센스를 따름